+++ helpers/ensureGuildConfig.js
@@ -1,3 +1,9 @@
+/**
+ * (c) 2022 Sakuragasaki46
+ * 
+ * See LICENSE for license details
+ */
+
 const { MessageEmbed, Permissions } = require('discord.js');
 
 module.exports = async function(interaction, variables){