+++ helpers/money.js
@@ -1,3 +1,9 @@
+/**
+ * (c) 2022 Sakuragasaki46
+ * 
+ * See LICENSE for license details
+ */
+
 module.exports = function(v, symbol = 'ꝁ') {
   return `${symbol} ${(+v).toLocaleString('en-US')}`;
 }