+++ scheduled/420.js
@@ -1,9 +1,17 @@
+/**
+ * (c) 2022 Sakuragasaki46
+ * 
+ * See LICENSE for license details
+ */
 
+const { four20Channel } = require('../config.json');
 
 module.exports = {
   time: {hour: 4, minute: 20, tz: 'Europe/Rome'},
   async execute(client) {
-    const channel = client.channels.cache.get('880805122710593617');
+    if (!four20Channel) return;
+
+    const channel = client.channels.cache.get(four20Channel);
     if (channel) {
       channel.send("4:20, ora di farsi una bella canna");
     } else {